Address scalastyle concerns
