+++ project/plugins.sbt
@@ -1,2 +1,4 @@
 logLevel := Level.Warn
 addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.5.1")
+addSbtPlugin("org.scalastyle" %% "scalastyle-sbt-plugin" % "1.0.0")
+
