+++ src/main/scala-2.12/com/github/nmdguerreiro/opencage/geocoder/OpenCageClient.scala
@@ -2,24 +2,31 @@ package com.github.nmdguerreiro.opencage.geocoder
 
 import com.github.nmdguerreiro.opencage.geocoder.OpenCageClient.Scheme.Scheme
 import com.softwaremill.sttp._
+import com.softwaremill.sttp.StatusCodes
 import com.softwaremill.sttp.asynchttpclient.future.AsyncHttpClientFutureBackend
 import com.softwaremill.sttp.circe._
 import io.circe.parser._
-import org.asynchttpclient.{DefaultAsyncHttpClientConfig, AsyncHttpClientConfig}
 
 import scala.concurrent.{ExecutionContext, Future}
+import scala.language.postfixOps
 import scala.util.{Failure, Success, Try}
 
 /**
  * Creates a new client for the OpenCage forward and reverse geocoders.
  *
- * The client auto-closes its resources if used within a try-with-resources block. Otherwise, you should call close when you're done with it to avoid resource leakage.
+ * The client auto-closes its resources if used within a try-with-resources block. Otherwise, you should call close
+ * when you're done with it to avoid resource leakage.
  *
  * @param authKey The auth key provided to you by OpenCage
  * @param scheme http/https
- * @param hostname Defaults to com.github.nmdguerreiro.opencage.geocoder.OpenCageClient#defaultHostname(), but can be overridden for testing purposes
- * @param backend The AHC backend is used by default. If you need to configure proxies, timeouts, connection pools, etc, you can do that yourself, by providing a custom backend. That can easily be done
- *                by calling com.softwaremill.sttp.asynchttpclient.future.AsyncHttpClientFutureBackend#apply(com.softwaremill.sttp.SttpBackendOptions, scala.concurrent.ExecutionContext) and/or by supplying it a custom configuration (com.softwaremill.sttp.asynchttpclient.future.AsyncHttpClientFutureBackend#usingConfig(org.asynchttpclient.AsyncHttpClientConfig, scala.concurrent.ExecutionContext)) before calling this constructor.
+ * @param hostname Defaults to com.github.nmdguerreiro.opencage.geocoder.OpenCageClient#defaultHostname(), but can be
+  *                 overridden for testing purposes
+ * @param backend The AHC backend is used by default. If you need to configure proxies, timeouts, connection pools,
+ *                etc, you can do that yourself, by providing a custom backend. That can easily be done
+ *                by calling com.softwaremill.sttp.asynchttpclient.future.AsyncHttpClientFutureBackend#apply(com.softwaremill.sttp.SttpBackendOptions,
+ *                scala.concurrent.ExecutionContext) and/or by supplying it a custom configuration
+ *                (com.softwaremill.sttp.asynchttpclient.future.AsyncHttpClientFutureBackend#usingConfig(org.asynchttpclient.AsyncHttpClientConfig,
+ *                scala.concurrent.ExecutionContext)) before calling this constructor.
  */
 class OpenCageClient(authKey: String,
                      scheme: Scheme = OpenCageClient.defaultScheme,
@@ -68,7 +75,7 @@ class OpenCageClient(authKey: String,
 
   private def buildUri(query: String, withAnnotations: Boolean): Uri = {
 
-    val uri = (uri"${scheme.toString}://$hostname:$port/geocode/v1/json").params(
+    val uri = uri"${scheme.toString}://$hostname:$port/geocode/v1/json".params(
       Map(
         "q" -> query,
         "key" -> authKey
@@ -80,9 +87,10 @@ class OpenCageClient(authKey: String,
   }
 
   /**
-   * The client auto-closes its resources if used within a try-with-resources block. Otherwise, you should call close when you're done with it to avoid resource leakage.
+   * The client auto-closes its resources if used within a try-with-resources block. Otherwise, you should call
+   * close when you're done with it to avoid resource leakage.
    */
-  def close() = {
+  def close(): Unit = {
     backend.close()
   }
 
@@ -93,25 +101,25 @@ class OpenCageClient(authKey: String,
    * This method handles valid HTTP responses, and not timeouts, connection resets, etc (i.e. transport exception).
    */
   private def handleHttpRequestSuccess(httpResponse: Response[Either[io.circe.Error, OpenCageResponse]]): Try[OpenCageResponse] = {
-    val responseBody = httpResponse.body
+    def handleErrorResponse(errorResponse: OpenCageResponse) = {
+      httpResponse.code match {
+        case StatusCodes.BadRequest => Failure(new InvalidRequestError(s"Invalid request: ${errorResponse.status.message}"))
+        case StatusCodes.PaymentRequired => Failure(new QuotaExceededError(s"Quota exceeded: ${errorResponse.status.message}"))
+        case StatusCodes.Forbidden => Failure(new ForbiddenError(s"Forbidden: ${errorResponse.status.message}"))
+        case StatusCodes.RequestTimeout => Failure(new TimeoutError(s"Timeout: ${errorResponse.status.message}"))
+        case StatusCodes.Gone => Failure(new RequestTooLongError(s"Request too long: ${errorResponse.status.message}"))
+        case StatusCodes.TooManyRequests => Failure(new RateLimitExceededError(
+          s"Request too long: ${errorResponse.status.message}", errorResponse.rate.map(_.reset)))
+        case _ => Failure(new UnexpectedError(s"Unexpected error: ${errorResponse.status.message}"))
+      }
+    }
 
     httpResponse.body match {
       case Left(rawResp) => {
         // sttp sadly doesn't deserialize error responses, so we'll have to try it ourselves
         decode[OpenCageResponse](rawResp) match {
           case Left(err) => Failure(new DeserializationError(s"Failed to deserialize the response from the OpenCage server: $rawResp", err))
-          case Right(errorResponse) => {
-
-            httpResponse.code match {
-              case 400 => Failure(new InvalidRequestError(s"Invalid request: ${errorResponse.status.message}"))
-              case 402 => Failure(new QuotaExceededError(s"Quota exceeded: ${errorResponse.status.message}"))
-              case 403 => Failure(new ForbiddenError(s"Forbidden: ${errorResponse.status.message}"))
-              case 408 => Failure(new TimeoutError(s"Timeout: ${errorResponse.status.message}"))
-              case 410 => Failure(new RequestTooLongError(s"Request too long: ${errorResponse.status.message}"))
-              case 429 => Failure(new RateLimitExceededError(s"Request too long: ${errorResponse.status.message}", errorResponse.rate.map(_.reset)))
-              case _ => Failure(new UnexpectedError(s"Unexpected error: ${errorResponse.status.message}"))
-            }
-          }
+          case Right(errorResponse) => handleErrorResponse(errorResponse)
         }
       }
       case Right(eitherErrorOrResponse) => {
@@ -126,7 +134,7 @@ class OpenCageClient(authKey: String,
 
   private def buildUri(query: String, params: OpenCageClientParams): Uri = {
     // URI with default parameters
-    val uri = (uri"${scheme.toString}://$hostname:$port/geocode/v1/json").params(
+    val uri = uri"${scheme.toString}://$hostname:$port/geocode/v1/json".params(
       Map(
         "q" -> query,
         "key" -> authKey
@@ -140,7 +148,7 @@ class OpenCageClient(authKey: String,
     params.minConfidence.foreach { m => extraParams += ("min_confidence" -> m.toString) }
     params.limit.foreach { l => extraParams += ("limit" -> l.toString) }
 
-    if (!params.countryCodes.isEmpty) extraParams += ("countrycode" -> params.countryCodes.mkString(","))
+    if (params.countryCodes.nonEmpty) extraParams += ("countrycode" -> params.countryCodes.mkString(","))
 
     if (params.abbreviate) extraParams += ("abbrv" -> "1")
     if (params.withoutAnnotations) extraParams += ("no_annotations" -> "1")
